requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.11
pydantic==2.10.0
python-dotenv==1.0.1
langgraph==0.2.70
